Extract each custom component, shared types to their own files
