+++ client/wildcard/src/components/Form/MultiSelect/index.ts
@@ -1,1 +1,2 @@
 export * from './MultiSelect'
+export * from './types'
